April 12 5th

+++ src/test/java/Factile_Mode/DGBP_HSB_TestCase_15.java
@@ -76,8 +76,16 @@ public class DGBP_HSB_TestCase_15 extends Base {
 
 		driver1.quit();
 		driver.switchTo().window(driver.getWindowHandle());
-		driver.quit();
-		Runtime.getRuntime().exec("taskkill /F /IM chromedriver.exe /T");
+		driver.quit();String osName = System.getProperty("os.name");
+		if (osName.equals("Windows 10")) {
+			Runtime.getRuntime().exec("taskkill /F /IM chromedriver.exe /T");
+			}
+			else
+			{
+				String[] cmd = new String[]{"/bin/sh", "killchrome.sh"};
+				Process pr = Runtime.getRuntime().exec(cmd);
+			}
+
 	}
 
 	public void modeaterscreen() throws InterruptedException {
