+++ src/test/java/Factile_Mode/DGBP_HSB_TestCase_26.java
@@ -85,7 +85,16 @@ public class DGBP_HSB_TestCase_26 extends Base {
 		driver2.quit();
 		driver3.switchTo().window(driver3.getWindowHandle());
 		driver3.quit();
-		Runtime.getRuntime().exec("taskkill /F /IM chromedriver.exe /T");
+		String osName = System.getProperty("os.name");
+		if (osName.equals("Windows 10")) {
+				Runtime.getRuntime().exec("taskkill /F /IM chromedriver.exe /T");
+				}
+				else
+				{
+					String[] cmd = new String[]{"/bin/sh", "killchrome.sh"};
+					Process pr = Runtime.getRuntime().exec(cmd);
+				}
+
 	}
 
 	public void modeaterscreen() throws InterruptedException {
