+++ src/test/java/Factile_Mode/DGBP_HSB_TestCase_38.java
@@ -85,7 +85,16 @@ public class DGBP_HSB_TestCase_38 extends Base {
 	@AfterTest
 	public void tearDown() throws InterruptedException, IOException {
 
-		Runtime.getRuntime().exec("taskkill /F /IM chromedriver.exe /T");
+		String osName = System.getProperty("os.name");
+		if (osName.equals("Windows 10")) {
+				Runtime.getRuntime().exec("taskkill /F /IM chromedriver.exe /T");
+				}
+				else
+				{
+					String[] cmd = new String[]{"/bin/sh", "killchrome.sh"};
+					Process pr = Runtime.getRuntime().exec(cmd);
+				}
+
 	//	 driver1.quit(); driver.switchTo().window(driver.getWindowHandle());
 		//  driver.quit();
