+++ src/test/java/Factile_Mode/DGBP_HSB_TestCase_53.java
@@ -82,7 +82,15 @@ public class DGBP_HSB_TestCase_53 extends Base {
 		driver.quit();
 		
 		 driver2.switchTo().window(driver2.getWindowHandle()); driver2.quit();
-		 Runtime.getRuntime().exec("taskkill /F /IM chromedriver.exe /T");
+		 String osName = System.getProperty("os.name");
+		 if (osName.equals("Windows 10")) {
+		 		Runtime.getRuntime().exec("taskkill /F /IM chromedriver.exe /T");
+		 		}
+		 		else
+		 		{
+		 			String[] cmd = new String[]{"/bin/sh", "killchrome.sh"};
+		 			Process pr = Runtime.getRuntime().exec(cmd);
+		 		}
 
 	}
